alx_travel_app/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
celery>=5.3.0
redis>=4.5.0
python-decouple>=3.8
